requirements.txt
pyyaml
pandas
requests
openai
fastparquet